soccer-rig-server/requirements.txt
# Soccer Rig Server - Python Dependencies

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Database drivers (install one based on your database)
psycopg2-binary>=2.9.0  # PostgreSQL
# pymysql>=1.1.0  # MySQL (uncomment if using MySQL)

# File handling
python-magic>=0.4.27
watchdog>=3.0.0

# Video processing
ffmpeg-python>=0.2.0

# ML/Computer Vision (for player tracking and action detection)
opencv-python-headless>=4.8.0
numpy>=1.24.0
ultralytics>=8.0.0  # YOLO for player detection

# Task queue (for async processing)
celery>=5.3.0
redis>=5.0.0

# HTTP client
requests>=2.31.0
httpx>=0.25.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.0
humanize>=4.9.0

# Logging
structlog>=24.0.0
